requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-bootstrap5==23.3
qrcode==7.4.2
pillow==10.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
